Refactor user.js
