+++ src/styles.scss
@@ -1,57 +1,3 @@
-// @import "bootstrap";
-
-// body {
-//   width: 100%;
-//   padding: 0;
-//   margin: 0;
-// }
-
-// #content {
-//   display: flex;
-//   flex-direction: column;
-//   justify-content: center;
-//   align-items: center;
-//   color: white;
-//   height: 100vh;
-//   border: 3px solid black;
-// }
-
-// span,
-// input {
-//   border: none;
-//   border-radius: 3px;
-// }
-
-// form input {
-//   height: 35px;
-// }
-
-// #content span {
-//   background: #000;
-//   border-radius: 5px;
-//   margin-left: -4px;
-// }
-
-// #content div {
-//   align-self: center;
-// }
-
-// .remove-padding {
-//   padding-left: 0;
-//   padding-right: 0;
-// }
-
-// .selected {
-//   text-decoration: line-through;
-//   // color: transparent;
-//   text-shadow: 0 0 8px #444;
-// }
-
-// li {
-//   list-style: none;
-//   width: fit-content;
-// }
-
 @import "bootstrap";
 
 .active {
@@ -61,10 +7,6 @@
 #content {
   height: 100vh;
 }
-
-// #hidden {
-//   display: none;
-// }
 
 footer {
   position: fixed;
@@ -113,4 +55,9 @@ textarea {
 .done-task {
   text-decoration: line-through;
   text-shadow: 0 0 8px #444;
+}
+
+.project-list h6:hover {
+  background-color: lightgrey;
+  cursor: pointer;
 }
